feat(dalat): Implemented webshare methods in dalat
