+++ packages/dalat/lib/src/converters.dart
@@ -1,5 +1,6 @@
 import 'dart:typed_data';
 
+import 'package:dalat/dalat.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 typedef DeviceType = String;
