+++ packages/dalat/lib/src/instance.dart
@@ -3,6 +3,9 @@ import 'dart:ffi';
 import 'dart:isolate';
 
 import 'package:dalat/dalat.dart';
+import 'package:dalat/src/config.dart';
+import 'package:dalat/src/helpers.dart';
+import 'package:dalat/src/webshare.dart';
 import 'package:ffi/ffi.dart';
 
 import 'bindings.dart';
@@ -13,7 +16,9 @@ typedef PairRequestHandler = Future<PairResponse> Function(PairRequest);
 ///
 /// This class encapsulates the FFI handle management and provides a clean,
 /// Dart-idiomatic interface to the underlying Go implementation.
-class AlatInstance {
+class AlatInstance
+    with InstanceHelpers, InstanceWebShare, InstancePair, InstanceConfig {
+  @override
   final int handle;
 
   AlatInstance._(this.handle);
@@ -36,7 +41,7 @@ class AlatInstance {
       );
       if (handle <= 0) {
         throw Exception(
-          'Failed to create handle, got id: $handle; ${getAlatError()}',
+          'Failed to create handle, got id: $handle; ${InstanceHelpers.getAlatError()}',
         );
       }
       return AlatInstance._(handle);
@@ -86,36 +91,8 @@ class AlatInstance {
     }
   }
 
-  static String getAlatError() {
-    final msgPointer = bindings.get_error();
-    final error = msgPointer == nullptr
-        ? "Unknown error"
-        : msgPointer.cast<Utf8>().toDartString();
-    bindings.free_string(msgPointer);
-    return error;
-  }
-
   void stop() {
     bindings.stop_instance(handle);
-  }
-
-  void registerPairRequestHandler(PairRequestHandler fun) {
-    final nativeCallback =
-        NativeCallable<
-          Void Function(Int, Pointer<Char>, Pointer<Char>, Pointer<Char>)
-        >.listener(_pairRequestHandler);
-    bindings.register_async_pair_request_callback(
-      handle,
-      nativeCallback.nativeFunction,
-    );
-    _pairRequestHandlers[handle] = fun;
-    _nativeCallables[handle] = nativeCallback;
-
-    // Register the native callback with Go.
-    bindings.register_async_pair_request_callback(
-      handle,
-      nativeCallback.nativeFunction,
-    );
   }
 
   void dispose() {
@@ -123,125 +100,8 @@ class AlatInstance {
     unregisterPairRequestHandler();
   }
 
-  Future<AppConfig> getAppConfig() async {
-    return _jsonHelper(bindings.get_app_config_json, AppConfig.fromJson);
-  }
-
-  Future<void> setAppConfig(AppConfig settings) async {
-    return _jsonSetterHelper(settings.toJson(), bindings.set_app_config_json);
-  }
-
-  Future<ServiceConfig> getServiceConfig() {
-    return _jsonHelper(
-      bindings.get_service_config_json,
-      ServiceConfig.fromJson,
-    );
-  }
-
-  Future<void> setServiceConfig(ServiceConfig settings) {
-    return _jsonSetterHelper(
-      settings.toJson(),
-      bindings.set_service_config_json,
-    );
-  }
-
-  Future<List<FoundDevice>> getFoundDevices() {
-    return _jsonListHelper(
-      bindings.get_found_devices_json,
-      FoundDevice.fromJson,
-    );
-  }
-
-  Future<List<PairedDevice>> getPairedDevices() {
-    return _jsonListHelper(
-      bindings.get_paired_devices_json,
-      PairedDevice.fromJson,
-    );
-  }
-
-  Future<List<ConnectedDevice>> getConnectedDevices() {
-    return _jsonListHelper(
-      bindings.get_connected_devices_json,
-      ConnectedDevice.fromJson,
-    );
-  }
-
   Future<NodeStatus> getNodeStatus() {
-    return _jsonHelper(bindings.get_node_status_json, NodeStatus.fromJson);
-  }
-
-  Future<List<DeviceColor>> getAlatColors() {
-    final ptr = bindings.get_alat_device_colors_json();
-    if (ptr == nullptr) {
-      return Future.value([]);
-    }
-    try {
-      final jsonStr = ptr.cast<Utf8>().toDartString();
-      final List<dynamic> decoded = jsonDecode(jsonStr);
-      return Future.value(
-        decoded
-            .map((item) => DeviceColor.fromJson(item as Map<String, dynamic>))
-            .toList(),
-      );
-    } finally {
-      bindings.free_string(ptr);
-    }
-  }
-
-  Future<T> _jsonHelper<T>(
-    Pointer<Char> Function(int) ffiFunc,
-    T Function(Map<String, dynamic>) fromJson,
-  ) async {
-    final ptr = ffiFunc(handle);
-    if (ptr == nullptr) {
-      throw Exception(
-        'Failed to get data from Go core: function returned null pointer. ${getAlatError()}',
-      );
-    }
-    try {
-      final jsonStr = ptr.cast<Utf8>().toDartString();
-      return fromJson(jsonDecode(jsonStr));
-    } finally {
-      bindings.free_string(ptr);
-    }
-  }
-
-  Future<List<T>> _jsonListHelper<T>(
-    Pointer<Char> Function(int) ffiFunc,
-    T Function(Map<String, dynamic>) fromJson,
-  ) async {
-    final ptr = ffiFunc(handle);
-    if (ptr == nullptr) {
-      // An empty list is represented by a null pointer in this API
-      return [];
-    }
-    try {
-      final jsonStr = ptr.cast<Utf8>().toDartString();
-      final List<dynamic> decoded = jsonDecode(jsonStr) ?? [];
-      return decoded
-          .map((item) => fromJson(item as Map<String, dynamic>))
-          .toList();
-    } finally {
-      bindings.free_string(ptr);
-    }
-  }
-
-  Future<void> _jsonSetterHelper(
-    Map<String, dynamic> jsonData,
-    int Function(int, Pointer<Char>) ffiFunc,
-  ) async {
-    final jsonStr = jsonEncode(jsonData);
-    final jsonStrC = jsonStr.toNativeUtf8();
-    try {
-      final result = ffiFunc(handle, jsonStrC.cast());
-      if (result != 0) {
-        throw Exception(
-          'Failed to set data in Go core. Code: $result ${getAlatError()}',
-        );
-      }
-    } finally {
-      malloc.free(jsonStrC);
-    }
+    return jsonGetterHelper(bindings.get_node_status_json, NodeStatus.fromJson);
   }
 
   Future<SysInfo> queryConnectedDeviceSysInfo(String deviceId) {
@@ -253,42 +113,13 @@ class AlatInstance {
           deviceIdC.cast(),
         );
         if (infoC == nullptr) {
-          throw "Failed getting system information: ${getAlatError()}";
+          throw "Failed getting system information: ${InstanceHelpers.getAlatError()}";
         }
         try {
           final info = infoC.cast<Utf8>().toDartString();
           return SysInfo.fromJson(jsonDecode(info));
         } finally {
           bindings.free_string(infoC);
-        }
-      } finally {
-        malloc.free(deviceIdC);
-      }
-    });
-  }
-
-  Future<RequestPairResponse> requestPair(String deviceId) {
-    return Isolate.run(() {
-      final deviceIdC = deviceId.toNativeUtf8();
-      try {
-        final ptr = bindings.request_pair_found_device(
-          handle,
-          deviceIdC.cast(),
-        );
-        if (ptr == nullptr) {
-          return RequestPairResponse(
-            status: -1,
-            error: "Alat sent no reponse",
-            accepted: false,
-            reason: "Could not query device",
-          );
-        } else {
-          try {
-            final result = ptr.cast<Utf8>().toDartString();
-            return RequestPairResponse.fromJson(jsonDecode(result));
-          } finally {
-            bindings.free_string(ptr);
-          }
         }
       } finally {
         malloc.free(deviceIdC);
@@ -307,7 +138,7 @@ class AlatInstance {
           filesJsonC.cast(),
         );
         if (response < 0) {
-          throw "Error sending files: ${getAlatError()}";
+          throw "Error sending files: ${InstanceHelpers.getAlatError()}";
         }
       } finally {
         malloc.free(deviceIdC);
@@ -316,91 +147,10 @@ class AlatInstance {
     });
   }
 
-  static AppConfig createAppConfig() {
-    final ptr = bindings.default_app_config();
-    if (ptr == nullptr) {
-      throw "Could not create default app settings, backend sent invalid null response";
-    }
-    try {
-      final jsonStr = ptr.cast<Utf8>().toDartString();
-      final Map<String, dynamic> decoded = jsonDecode(jsonStr);
-      return AppConfig.fromJson(decoded);
-    } finally {
-      bindings.free_string(ptr);
-    }
-  }
-
-  static ServiceConfig createServiceConfig() {
-    final ptr = bindings.default_service_config();
-    if (ptr == nullptr) {
-      throw "Could not create default service configuration, backend sent invalid null response";
-    }
-    try {
-      final jsonStr = ptr.cast<Utf8>().toDartString();
-      final Map<String, dynamic> decoded = jsonDecode(jsonStr);
-      return ServiceConfig.fromJson(decoded);
-    } finally {
-      bindings.free_string(ptr);
-    }
-  }
-
-  void unregisterPairRequestHandler() {
-    _nativeCallables[handle]?.close();
-    _nativeCallables.remove(handle);
-    _pairRequestHandlers.remove(handle);
-  }
-
   Future<FileTransfersStatus> getFileTransfersStatus() {
-    return _jsonHelper(
+    return jsonGetterHelper(
       bindings.get_file_transfers_status,
       FileTransfersStatus.fromJson,
     );
-  }
-}
-
-final Map<int, PairRequestHandler> _pairRequestHandlers = {};
-final Map<int, NativeCallable> _nativeCallables = {};
-void _pairRequestHandler(
-  int handle,
-  Pointer<Char> requestIdC,
-  Pointer<Char> pairTokenC,
-  Pointer<Char> deviceDetailsC,
-) {
-  final handler = _pairRequestHandlers[handle];
-  try {
-    if (handler == null) return;
-    final requestId = requestIdC.cast<Utf8>().toDartString();
-    final pairToken = Uint8ListConverter().fromJson(
-      jsonDecode(pairTokenC.cast<Utf8>().toDartString()),
-    );
-    final deviceDetails = DeviceDetails.fromJson(
-      jsonDecode(deviceDetailsC.cast<Utf8>().toDartString()),
-    );
-
-    handler(
-      PairRequest(
-        requestid: requestId,
-        token: pairToken,
-        device: deviceDetails,
-      ),
-    ).then((response) {
-      final newRequestIdC = requestId.toNativeUtf8();
-      final reasonC = response.reason.toNativeUtf8();
-      try {
-        bindings.submit_pair_response(
-          handle,
-          newRequestIdC.cast(),
-          response.accepted,
-          reasonC.cast(),
-        );
-      } finally {
-        malloc.free(newRequestIdC);
-        malloc.free(reasonC);
-      }
-    });
-  } finally {
-    malloc.free(requestIdC);
-    malloc.free(pairTokenC);
-    malloc.free(deviceDetailsC);
   }
 }
